BREAKING CHANGE: remove hasUkisLayerMetadata and move addUkisLayerMetadata and getUkisLayerMetadata
